buffer some channels
without this buffering we hang forever if we try to send to a channel
_during_ a select.

Unbuffered channels block when written to until they are read. If you
write to an unbuffered channel in a select you will never get back to
the top of the select in order to read from where you wrote. Buffering
these two fixes the issue. The other channels can remain unbuffered
since they are written to from other goroutines.